moins de leak

+++ tools/lua2/Ref.cpp
@@ -69,7 +69,7 @@ namespace Tools { namespace Lua {
         auto itEnd = args.end();
         for (; it != itEnd; ++it)
             it->ToStack();
-        if (lua_pcall(this->_i, args.size(), LUA_MULTRET, 0))
+        if (lua_pcall(this->_i, static_cast<int>(args.size()), LUA_MULTRET, 0))
         {
             std::string e = "Lua::Ref: Error in function call: ";
             e += lua_tostring(this->_i, -1);
